Add files via upload
The project involves designing a complete database management system to address a practical database need and implementing a relational database based on that design. 
This project is designed to perform general information management tasks such as systematic collection, update, and retrieval of information for a small organization.
It includes a menu bar with features like edit record, delete record and add a new participant.